requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
build==0.7.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
distlib==0.3.4
distro==1.7.0
filelock==3.7.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
mozdownload==1.26.0
mozfile==2.1.0
mozinfo==1.2.2
numpy==1.21.6
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pep517==0.12.0
platformdirs==2.5.2
platinfo==0.15.0
progressbar2==4.0.0
py-firefox-driver-manager==0.0.4
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-utils==3.3.0
pytz==2022.1
pywin32-ctypes==0.2.0
redo==2.0.3
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
style==1.1.0
tomli==2.0.1
treeherder-client==5.0.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
update==0.0.1
urllib3==1.26.9
virtualenv==20.14.1
wsproto==1.1.0
zipp==3.8.0
